requirements.txt
aiodns==3.0.0
aiogram==2.20
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
cryptography==37.0.2
Deprecated==1.2.13
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
html5lib==1.1
humanfriendly==10.0
idna==3.3
install==1.3.5
kombu==5.2.4
lxml==4.8.0
multidict==6.0.2
numpy==1.22.3
outcome==1.1.0
packaging==21.3
peewee==3.14.10
pefile==2021.9.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycares==4.1.2
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyVirtualDisplay==3.0
redis==4.2.2
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.7.0
webencodings==0.5.1
wrapt==1.14.0
wsproto==1.1.0
yarl==1.7.2
